Change `int` to `long` for Platform Compatibility

+++ YookaiC/Game/SDLEx/Vulkan/SDLExVulkan.h
@@ -70,9 +70,9 @@ VkImageCreateInfo create_image(unsigned width, unsigned height, VkFormat format,
 VkImageView create_image_view(VkImage image, VkFormat format);
 void transition_image_layout(VkImage image, VkImageLayout oldLayout, VkImageLayout newLayout);
 void copy_buffer_to_image(VkBuffer buffer, VkImage image, unsigned width, unsigned height);
-int load_texture2d(const char * filename);
-void dispose_texture2d(int texture_id);
-void bind_texture2d(unsigned imageIndex, int texture_id);
+long load_texture2d(const char * filename);
+void dispose_texture2d(long texture_id);
+void bind_texture2d(unsigned imageIndex, long texture_id);
 
 void create_descriptor_pool();
 void create_descriptor_sets();
